+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- File selection step when interacting with videos, animations, RAWs 
+  and other entries with multiple files. Now the original file is selected automatically
+
 ## [1.32.0] - 2024-11-11
 
 ### Added